+++ apps/engine/src/constants.ts
@@ -1,0 +1,4 @@
+import path from 'node:path';
+
+export const DEFAULT_ENGINE_PLUGIN_PATH = path.join('plugins', 'node_modules');
+export const DEFAULT_WEB_PACKAGE_NAME = '@tachybase/app-web';
