+++ apps/engine/src/defaultModules.ts
@@ -1,0 +1,24 @@
+export const defaultModules = [
+  'acl',
+  'app-info',
+  'auth',
+  'backup',
+  'cloud-component',
+  'collection',
+  'cron',
+  'data-source',
+  'error-handler',
+  'event-source',
+  'file',
+  'message',
+  'pdf',
+  'ui-schema',
+  'user',
+  'web',
+  'worker-thread',
+  'instrumentation',
+  'workflow',
+  'env-secrets',
+  'hera', // FIXME: refactor later
+  'multi-app',
+];
