+++ apps/engine/src/defaultPlugins.ts
@@ -1,0 +1,24 @@
+export const defaultPlugins = [
+  'action-bulk-edit',
+  'action-bulk-update',
+  'action-custom-request',
+  'action-duplicate',
+  'action-export',
+  'action-import',
+  'action-print',
+  'block-calendar',
+  'block-charts',
+  'block-gantt',
+  'block-kanban',
+  'block-presentation',
+  'field-china-region',
+  'field-formula',
+  'field-sequence',
+  'field-encryption',
+  'log-viewer',
+  'otp',
+  'full-text-search',
+  'password-policy',
+  'auth-pages',
+  'manual-notification',
+];
