+++ apps/engine/src/utils.ts
@@ -9,62 +9,11 @@ import { mkdir, unlink } from 'node:fs/promises';
 import { dirname, join, resolve } from 'node:path';
 import { pipeline } from 'node:stream/promises';
 
-import yoctoSpinner from '@socketregistry/yocto-spinner/index.cjs';
 import { config } from 'dotenv';
 import npmRegistryFetch from 'npm-registry-fetch';
 import * as tar from 'tar';
 
-const modules = [
-  'acl',
-  'app-info',
-  'auth',
-  'backup',
-  'cloud-component',
-  'collection',
-  'cron',
-  'data-source',
-  'error-handler',
-  'event-source',
-  'file',
-  'message',
-  'pdf',
-  'ui-schema',
-  'user',
-  'web',
-  'worker-thread',
-  'instrumentation',
-  'workflow',
-  'env-secrets',
-  'hera', // FIXME: refactor later
-  'multi-app',
-];
-
-const defaultPlugins = [
-  'action-bulk-edit',
-  'action-bulk-update',
-  'action-custom-request',
-  'action-duplicate',
-  'action-export',
-  'action-import',
-  'action-print',
-  'block-calendar',
-  'block-charts',
-  'block-gantt',
-  'block-kanban',
-  'block-presentation',
-  'field-china-region',
-  'field-formula',
-  'field-sequence',
-  'field-encryption',
-  'log-viewer',
-  'otp',
-  'full-text-search',
-  'password-policy',
-  'auth-pages',
-  'manual-notification',
-];
-
-function initEnvFile(name: string) {
+export function initEnvFile(name: string) {
   const envPath = resolve(name, '.env');
   if (!fs.existsSync(envPath)) {
     fs.copyFileSync(resolve(__dirname, '../presets/.env.example'), envPath);
@@ -196,43 +145,4 @@ export async function downloadTar(packageName: string, target: string) {
   });
 
   await unlink(tarballFile);
-}
-
-export async function prepare(name: string, plugins = defaultPlugins) {
-  if (fs.existsSync(name)) {
-    console.log(`project folder ${name} already exists, exit now.`);
-    return;
-  }
-  fs.mkdirSync(name);
-  initEnvFile(name);
-
-  const prefix = `${name}/plugins/node_modules`;
-  // 安装前端代码
-  console.log('🚀 ~ start download ~ front end files');
-  const spinner = yoctoSpinner({ text: `Loading @tachybase/app-web` }).start();
-  await downloadTar('@tachybase/app-web', `${prefix}/@tachybase/app-web`);
-  spinner.success();
-  console.log();
-
-  console.log('🚀 ~ start download ~ required modules');
-  // 安装必须得模块
-  const moduleNames = modules.map((moduleName) => `@tachybase/module-${moduleName}`);
-  let index = 1;
-  for (const moduleName of moduleNames) {
-    const spinner = yoctoSpinner({ text: `[${index++}/${moduleNames.length}] Loading ${moduleName}` }).start();
-    await downloadTar(moduleName, `${prefix}/${moduleName}`);
-    spinner.success();
-  }
-  console.log();
-
-  console.log('🚀 ~ start download ~ plugins');
-  // 安装可选的模块，由参数指定
-  index = 1;
-  const pluginNames = plugins.map((pluginName: string) => `@tachybase/plugin-${pluginName}`);
-  for (const pluginName of pluginNames) {
-    const spinner = yoctoSpinner({ text: `[${index++}/${pluginNames.length}] Loading ${pluginName}` }).start();
-    await downloadTar(pluginName, `${prefix}/${pluginName}`);
-    spinner.success();
-  }
-  console.log();
 }
